# tools/openapi.yaml
openapi: 3.0.3
info:
  version: 0.9.6
  title: LANraragi API
...
paths:
...
  /database/stats:
    get:
      operationId: getStatistics
      x-mojo-to: api-database#serve_tag_stats
      summary: Get Statistics
      description: Get tags from the database, with a value symbolizing their prevalence.
      tags:
        - database
      parameters:
        - name: minweight
          in: query
          required: false
          description: >-
            Add this parameter if you want to only get tags whose weight is at
            least the given minimum.  

            Default is 1 if not specified, to get all tags.
          schema:
            type: integer
        - name: hide_excluded_namespaces
          in: query
          required: false
          description: >-
            Set to true to exclude namespaces configured in server settings from
            the results. Default behavior returns all tags.
          schema:
            type: string
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                      - text
                  properties:
                    namespace:
                      type: string
                      nullable: true
                      description: Namespace of the tag
                    text:
                      type: string
                      description: The tag itself
                    weight:
                      type: integer
                      description: Weight of the tag in the database
              example:
                - namespace: character
                  text: jeanne alter
                  weight: 2
                - namespace: character
                  text: xuanzang
                  weight: 2
                - namespace: parody
                  text: fate grand order
                  weight: 3
                - namespace: null
                  text: artbook
                  weight: 3
  /database/backup:
    get:
      security:
        - api_key: []
      operationId: getBackupJson
      x-mojo-to: api-database#serve_backup
      summary: 🔑 Get a backup JSON
      description: >-
        Scans the entire database and returns a backup in JSON form.  
        Consider using the POST variant if your database is large, as this basic GET endpoint might time out if it takes too long to generate the backup.  

        This backup can be reimported manually through the Backup and Restore feature.
      tags:
        - database
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                type: object
                properties:
                  archives:
                    type: array
                    description: Array of archive metadata
                    items:
                      $ref: '#/components/schemas/BackupArchiveMetadataJson'
                  categories:
                    type: array
                    description: Array of category metadata
                    items:
                      $ref: '#/components/schemas/BackupCategoryMetadataJson'
                  tankoubons:
                    type: array
                    description: Array of tankoubon metadata
                    items:
                      $ref: '#/components/schemas/TankoubonMetadataJson'
                example:
                  archives:
                    - arcid: 28697b96f0ac5858be2614ed10ca47742c9522fd
                      title: Fate GO MEMO
                      tags: parody:fate grand order, group:wadamemo, artist:wada rco, artbook, full color, super:test, date_added:1553537258
                      summary: null
                      thumbhash: ec2a0ca3a3da67a9390889f0910fe494241faa9a
                      filename: FateGOMEMO
                    - arcid: d1858d5dc36925aa66be072a97817650d39de166
                      title: test title
                      tags: null
                      summary: null
                      thumbhash: 9846bb6d62949b56545c42e88d8010446b65702e
                      filename: (Memes)9B789D38B0784C5FBC9D59A9F24D20F2
                  categories:
                    - archives: []
                      catid: SET_1749925633
                      name: english
                      search: language:english
                  tankoubons:
                    - tankid: TANK_1749925516
                      name: sailor moon
                      archives:
                        - 28697b96f0ac5858be2614ed10ca47742c9522fd
                        - d1858d5dc36925aa66be072a97817650d39de166
    post:
      security:
        - api_key: []
      operationId: queueBackupJob
      x-mojo-to: api-database#queue_backup
      summary: 🔑 Queue a backup job
      description: >-
        Queues a Minion job to generate a backup JSON file.  

        Use the returned job ID to check progress and download the file once complete.
      tags:
        - database
      responses:
        '200':
          description: Enqueued job
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation:
                    type: string
                    enum: 
                      - queue_backup
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                  job:
                    type: integer
                example:
                  job: 42
                  operation: queue_backup
                  success: 1
  /database/backup/{jobid}:
    get:
      security:
        - api_key: []
      operationId: downloadBackup
      x-mojo-to: api-database#download_backup
      summary: 🔑 Download backup JSON from completed job
      description: >-
        Downloads the backup JSON file generated by a completed backup job.
      tags:
        - database
      parameters:
        - name: jobid
          in: path
          required: true
          description: ID of the completed backup job.
          schema:
            type: integer
        - name: format
          in: query
          required: false
          description: >-
            Format of the returned backup. Defaults to `file`. `json` will return the backup as a JSON response.
          schema:
            type: string
            enum:
              - json
              - file
      responses:
        '200':
          description: Backup JSON file
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Job not found or not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /database/restore:
    post:
      security:
        - api_key: []
      operationId: queueRestoreJob
      x-mojo-to: api-database#queue_restore
      summary: 🔑 Queue a restore job
      description: >-
        Queues a Minion job to restore from a backup JSON file.  

        Use the returned job ID to check progress.
      tags:
        - database
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Backup JSON file to restore
              required:
                - file
      responses:
        '200':
          description: Enqueued job
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation:
                    type: string
                    enum: 
                      - queue_restore
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                  job:
                    type: integer
                example:
                  job: 43
                  operation: queue_restore
                  success: 1
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /database/isnew:
    delete:
      security:
        - api_key: []
      operationId: clearNewAll
      x-mojo-to: api-database#clear_new_all
      summary: 🔑 Clear All "New" flags
      description: Clears the "New!" flag on all archives.
      tags:
        - database
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /database/drop:
    post:
      security:
        - api_key: []
      operationId: dropDatabase
      x-mojo-to: api-database#drop_database
      summary: 🔑 Drop the Database
      description: >-
        Delete the entire database, including user preferences.

        This is a rather dangerous endpoint, invoking it might lock you out of
        the server as a client!
      tags:
        - database
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /database/clean:
    post:
      security:
        - api_key: []
      operationId: cleanDatabase
      x-mojo-to: api-database#clean_database
      summary: 🔑 Clean the Database
      description: >-
        Cleans the Database, hiding then removing entries for files that are no longer on the filesystem.  

        Entries that are no longer on the filesystem are only unlinked at first so they don't appear in the UI -- A subsequent run of this cleanup will **delete** unlinked entries.
      tags:
        - database
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation:
                    type: string
                    description: Name of the operation
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                    description: Returns 1 if operation successful, else 0.
                  deleted:
                    type: integer
                    description: Amount of unlinked DB entries that were deleted
                  unlinked:
                    type: integer
                    description: Amount of live DB entries that got unlinked - Run Clean again to delete them. 
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: Authorization
      in: header
      description: 'Use Authorization: Bearer <base64(api_key)>'
  schemas:
...
    OperationResponse:
      type: object
      required:
        - operation
        - success
      properties:
        operation:
          type: string
          description: Name of operation
        error:
          type: string
          description: Error message if any
        successMessage:
          type: string
          description: Success message if any
        success:
          type: integer
          description: Returns 1 if operation was successful, else 0
          enum:
            - 0
            - 1
...
    TankoubonMetadataJson:
      type: object
      description: Tankoubon metadata
      required:
        - tankid
      properties:
        archives:
          type: array
          description: Array of archive IDs
          items:
            type: string
            minLength: 40
            maxLength: 40
        tankid:
          type: string
          description: ID of the tankoubon
        name:
          type: string
          description: Name of the tankoubon
      example:
        tankid: TANK_1749925516
        name: sailor moon
        archives:
          - e69e43e1355267f7d32a4f9b7f2fe108d2401ebf
          - e69e43e1355267f7d32a4f9b7f2fe108d2401ebg
    BackupArchiveMetadataJson:
      type: object
      description: JSON object for archive metadata in as returned in a backup
      required:
        - arcid
        - title
        - tags
        - summary
        - filename
      properties:
        arcid:
          type: string
          minLength: 40
          maxLength: 40
          description: Unique identifier for the archive
        title:
          type: string
          description: Title of the archive
        tags:
          type: string
          description: Comma-separated list of tags associated with the archive
        summary:
          type: string
          nullable: true
          description: Summary description of the archive
        thumbhash:
          type: string
          nullable: true
          minLength: 40
          maxLength: 40
          description: Thumbnail hash, or null if not set
        filename:
          type: string
          description: Filename of the archive
    BackupCategoryMetadataJson:
      type: object
      description: >-
        JSON object for category metadata in as returned in a backup
      required:
        - catid
      properties:
        archives:
          type: array
          description: Array of archive IDs
          items:
            type: string
            minLength: 40
            maxLength: 40
        catid:
          type: string
          description: ID of the category
          minLength: 14
          maxLength: 14
        name:
          type: string
          description: Name of the category
        search:
          type: string
          description: Category search filter
...
servers:
  - url: https://lrr.tvc-16.science/api